pyproject.toml
[tool.poetry]
name = "coronavirus"
version = "0.1.2"
description = "coronavirus"
authors = ["hulk"]

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[tool.poetry.dependencies]
python = "^3.7"
jinja2 = "2.11.2"
aiofiles = "0.6.0"
fastapi = "0.64.0"
email-validator = "1.1.2"
requests = "^2.25.1"
uvicorn = "^0.14.0"
python-multipart = "^0.0.5"
passlib = "^1.7.4"
python-jose = "3.3.0"
bcrypt = "^3.1.0"
tortoise-orm = "^0.17.6"
aiomysql = "^0.0.21"
aerich = "^0.5.7"
httpx = "^0.19.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
